Pencil SKetch/requirements.txt

numpy==1.21.6

Pillow==9.2.0
opencv-python-headless
streamlit==1.11.1
